tweaks: SQLAlchemy models modernized to supress warnings

+++ app/schemas/product.py
@@ -1,16 +1,18 @@
+from decimal import Decimal
+
 from pydantic import BaseModel, Field
 
 
 class ProductCreate(BaseModel):
     name: str
-    price: float = Field(gt=0)
+    price: Decimal = Field(gt=0)
     quantity: int = Field(ge=0, default=0)
 
 
 class ProductResponse(BaseModel):
     id: int
     name: str
-    price: float
+    price: Decimal
     quantity: int
 
     class Config:
@@ -19,5 +21,5 @@ class ProductResponse(BaseModel):
 
 class ProductUpdate(BaseModel):
     name: str
-    price: float = Field(gt=0)
+    price: Decimal = Field(gt=0)
     quantity: int = Field(ge=0, default=0)
